Update pulseeffects to 4.2.5
Summary:
Added a plugin that adapts the audio buffer size to a power of 2 value. We were using the block size property of the pulsesrc plugin before but it was not working well. Sometimes Pulseaudio was not able to give a buffer with the size we needed and some noises were created because of GStreamer buffer under flows. Now it should be possible to use higher block sizes without noises. A larger block may be interesting to users using weak processors as a few plugins use less CPU with larger block sizes. As always there is a trade-off. Large blocks increase latency.

Signed-off-by: Pierre-Yves <pyu@riseup.net>

Test Plan: Open a video in MPV and a stream in Firefox, visualize them in pulseeffect

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Differential Revision: https://dev.getsol.us/D3890